LPFG-1041: Added authorisation to module report by profession endpoints.

+++ src/main/java/uk/gov/cshr/report/service/LearningCatalogueService.java
@@ -36,7 +36,7 @@ public class LearningCatalogueService {
         return httpService.getMap(moduleUri, Module.class);
     }
 
-    public Map<String, Module> getModuleMap(int professionId) {
+    public Map<String, Module> getModuleMap(long professionId) {
         URI uri = URI.create(String.format(moduleByProfessionUri, professionId));
         return httpService.getMap(uri, Module.class);
     }
